if time (or byobu's "now") ever goes backwards, invalidate last-run cache
time could go backwards via a setting of system date backwards.
It could also go backwards if byobu-statusd was running, and then was
killed and started again, but that time it decided to use 'uptime' or
SECONDS to get its 'now'.

in the case of SECONDS, every byobu-statusd will end up on first run
invalidating all cache (which isn't really so bad)