Add support for tmux grouped sessions
The default tmux behaviour when connecting multiple clients is a little
unexpected. All of the clients are connected to the same view. When the
view on one client is changed, all the other client views get changed
also. With this behaviour it isn't possible to have a different view in
each client connection as one would have with multiple xterm windows.
This differs from the behaviour of screen where the viewport of each
client is independent.

Fix this by making use of the session group feature in tmux. With
session groups, each client creates a new session, but groups it with
an existing session so it can share windows.

However, when a client detaches from a session in a session group, the
session instance doesn't get removed. If one attaches and detaches from
tmux many times, a lot of unused sessions will accumulate (as seen with
the 'tmux list-sessions' command). Solve this by adding a "cull_zombies"
step when connecting to a session. When attaching to a session,
cull_zombies() checks if there are any detached group sessions that
should be killed.