bashrc: export BYOBU_RUN_DIR and BYOBU_PREFIX
This is to address https://launchpad.net/bugs/2051995 where users will
see the phrase

bash: /timer.6217: Permission denied

Whenever any command is ran. This occurs because byobu-prompt uses this
bashrc and tries to write to $BYOBU_RUN_DIR/timer.$$ but BYOBU_RUN_DIR
is not yet defined. This caused byobu to try to write to /timer.$$ which
fails as a regular user.

Also added BYOBU_PREFIX in-case the user sets BYOBU_DISTRO to Ubuntu.

Signed-off-by: Mitchell Dzurick <mitchell.dzurick@canonical.com>