use VERSION in configure.ac
The generated usr/bin/byobu did not track the version specified in
configure.ac (via the AC_INIT macro) which lead to an incorrect naming
of the distribution file when using "make dist".

This commit makes usr/bin/byobu.in reference the autoconf VERSION
variable (derived from the AC_INIT macro) to ensure that there is one
source of version information.